[#33] Re-implement the ItemRepository using JPA instead of JDBC
